requirements.txt
httpx>=0.25.0
aiogram>=3.1.1
python-dotenv>=1.0.0
py-clob-client>=0.1.0
websockets>=12.0
tzdata>=2024.1
